src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "pymongo>=4.12.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
